Fix #3616: Step plot fail on Windows with datetime data

+++ bokeh/charts/glyphs.py
@@ -440,11 +440,11 @@ class StepGlyph(LineGlyph):
         elif self.y is None:
             y = self.x.index
 
-        xs = np.empty(2*len(x)-1, dtype=np.int)
+        xs = np.empty(2*len(x)-1, dtype=x.dtype)
         xs[::2] = x[:]
         xs[1::2] = x[1:]
 
-        ys = np.empty(2*len(y)-1)
+        ys = np.empty(2*len(y)-1, dtype=y.dtype)
         ys[::2] = y[:]
         ys[1::2] = y[:-1]
